requirements.txt
python-docx==0.8.11
PyMuPDF==1.23.8
PyPDF2==3.0.1
